requirements.txt
ipywidgets
kornia 
adabelief-pytorch==0.2.0